Fix button component
refs #31
