+++ components/auth-buttons/SignOutButton.tsx
@@ -7,6 +7,5 @@ export default function SignOutButton() {
     return <Button onClick={() => signOut()}
         text='Sign Out'
         dataCy="signout-button"
-        color="primary"
     />
 }
