+++ components/buttons/Button.tsx
@@ -1,12 +1,6 @@
-import { ThemeColorPicker } from "@/theme/color";
-import { ColorSwatch } from "@/theme/props";
 import { MouseEventHandler } from "react";
 
 interface ButtonProps {
-    /**
-     * Button color
-     */
-    color?: ColorSwatch;
     /**
      * Button text
      */
@@ -22,11 +16,10 @@ interface ButtonProps {
 }
 
 export default function Button(props: ButtonProps) {
-    const color = ThemeColorPicker.backgroundColor(props.color);
     const cypressSelector = props.dataCy ?? "button";
 
     return (
-        <button className={`${color} rounded-full px-10 py-4 transition-all hover:shadow-md`}
+        <button className="bg-amber-400 rounded-full px-10 py-4 transition-all hover:shadow-md"
             data-cy={cypressSelector}
             onClick={props.onClick}>
             {props.text ?? "Button"}
