+++ components/header/Header.tsx
@@ -1,8 +1,6 @@
 import { headerFont } from "@/app/fonts";
-import { ThemeColorPicker } from "@/theme/color";
-import PropsTheme from "@/theme/props";
 
-interface HeaderProps extends PropsTheme {
+interface HeaderProps {
     /**
      * Text to display
      */
@@ -10,11 +8,9 @@ interface HeaderProps extends PropsTheme {
 }
 
 export default function Header(props: HeaderProps) {
-    const textColor = ThemeColorPicker.textColor(props.theme);
-
     return (
-        <header className={`${headerFont.className} ${textColor} text-2xl`} data-cy="header">
-            {props.value}
+        <header className={`${headerFont.className} text-gray-800 text-2xl`} data-cy="header">
+            {props.value ?? "Header"}
         </header>
     );
 }
