+++ stories/Header.stories.tsx
@@ -10,9 +10,4 @@ const meta: Meta<typeof Header> = {
 export default meta;
 type Story = StoryObj<typeof meta>;
 
-export const Example: Story = {
-  args: {
-    value: "Header",
-    theme: "dark"
-  },
-};
+export const Example: Story = {};
