--- theme/props.ts
@@ -1,6 +1,0 @@
-export type Theme = "dark" | "light";
-export type ColorSwatch = "primary" | "secondary";
-
-export default interface PropsTheme {
-  theme?: Theme;
-}
